back-end/requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
pyserial==3.5